Adding Resale-support option for Owner
1. Added REST API at OCS, which is called upon by OPS during Transfer
Ownership 2, Type 50, to know whether the RESALE protcol should be
supported for a particular device. By default, RESALE is supported.
2. To facilitate the above, adding a sample RESALE flag at OCS, which when
set to true, indicates that the Owner supports resale protocol for all
the devices. By default, RESALE is flag is set to true.
3. Fixing an issue where the credential-reuse flag was being processed
incorrectly, and setting the default value to false.

Signed-off-by: Chandrakar, Prateek <prateek.chandrakar@intel.com>