Fixing memory leaks from HttpClient
Creating a local single-thread ExecutorService and assigning to the
HttpClient so that HttpClient doesn't create its own. Also, manually
shutting down the Executor and nulling HttpClient so that they are
available for garbage collection.
Additionally, updating the pids limit of the docker container from 100
to 300 to allow thread creation by increased number of devices.

Signed-off-by: Chandrakar, Prateek <prateek.chandrakar@intel.com>