Adding default HTTP version to HTTPClient
When IOT Platform SDK is executed in an open network, the HTTP Clients
are not able to connect to hosted Rendezvous Service.

The issue is resolved by setting the HTTP version to 1.1 explicitly
while creating the required HTTP clients.

While at it, update the third-party package vesions to fix known
security vulnerabilities.

Signed-off-by: Chandrakar, Prateek <prateek.chandrakar@intel.com>
Signed-off-by: Behera, Tushar Ranjan <tushar.ranjan.behera@intel.com>